requirements_gpu.txt
Pillow>=9.3.0
streamlit>=1.21.0
numpy>=1.23.0
onnxruntime-gpu
torch==1.11.0
torchvision==0.12.0
